Extract aggregate row codec
